refactor: 쿠폰 관련 테스트 클래스들 통합
