--- backend/src/main/java/com/woowacourse/kkogkkog/application/dto/CouponsResponse.java
@@ -1,18 +1,0 @@
-package com.woowacourse.kkogkkog.application.dto;
-
-import java.util.List;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-// TODO: refactor to Response<T>
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-public class CouponsResponse {
-
-    private List<CouponResponse> data;
-
-    public CouponsResponse(List<CouponResponse> data) {
-        this.data = data;
-    }
-}
