+++ backend/src/test/java/com/woowacourse/kkogkkog/application/CouponServiceTest.java
@@ -4,6 +4,7 @@ import static com.woowacourse.kkogkkog.fixture.MemberFixture.NON_EXISTING_MEMBER
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import com.woowacourse.kkogkkog.application.dto.CouponMemberResponse;
 import com.woowacourse.kkogkkog.application.dto.CouponResponse;
 import com.woowacourse.kkogkkog.application.dto.CouponSaveRequest;
 import com.woowacourse.kkogkkog.domain.Member;
@@ -63,6 +64,68 @@ public class CouponServiceTest extends ServiceTest {
         void findById_notFound() {
             assertThatThrownBy(() -> couponService.findById(1L))
                     .isInstanceOf(CouponNotFoundException.class);
+        }
+    }
+
+    @DisplayName("사용자가 보낸 쿠폰들이 조회된다.")
+    @Nested
+    class FindBySenderTest {
+
+        @DisplayName("조회되는 쿠폰 개수 확인")
+        @Test
+        void couponCount() {
+            couponService.save(toCouponSaveRequest(ROOKIE, List.of(ARTHUR, JEONG, LEO)));
+            couponService.save(toCouponSaveRequest(JEONG, List.of(ARTHUR, LEO)));
+            List<CouponResponse> actual = couponService.findAllBySender(ROOKIE.getId());
+
+            assertThat(actual.size()).isEqualTo(3);
+        }
+
+        @DisplayName("조회되는 쿠폰의 보낸 사람 정보 확인")
+        @Test
+        void senderId() {
+            couponService.save(toCouponSaveRequest(ROOKIE, List.of(ARTHUR, JEONG, LEO)));
+            couponService.save(toCouponSaveRequest(JEONG, List.of(ARTHUR, LEO)));
+            Long senderId = ROOKIE.getId();
+
+            List<Long> actual = couponService.findAllBySender(ROOKIE.getId())
+                    .stream().map(CouponResponse::getSender)
+                    .map(CouponMemberResponse::getId)
+                    .collect(Collectors.toList());
+            List<Long> expected = List.of(senderId, senderId, senderId);
+
+            assertThat(actual).isEqualTo(expected);
+        }
+    }
+
+    @DisplayName("사용자가 받은 쿠폰들이 조회된다.")
+    @Nested
+    class FindByReceiverTest {
+
+        @DisplayName("조회되는 쿠폰 개수 확인")
+        @Test
+        void couponCount() {
+            couponService.save(toCouponSaveRequest(ARTHUR, List.of(JEONG, LEO)));
+            couponService.save(toCouponSaveRequest(LEO, List.of(JEONG, ARTHUR)));
+            List<CouponResponse> actual = couponService.findAllByReceiver(JEONG.getId());
+
+            assertThat(actual.size()).isEqualTo(2);
+        }
+
+        @DisplayName("조회되는 쿠폰의 받은 사람 정보 확인")
+        @Test
+        void receiverId() {
+            couponService.save(toCouponSaveRequest(ARTHUR, List.of(JEONG, LEO)));
+            couponService.save(toCouponSaveRequest(LEO, List.of(JEONG, ARTHUR)));
+            Long receiverId = JEONG.getId();
+
+            List<Long> actual = couponService.findAllByReceiver(receiverId)
+                    .stream().map(CouponResponse::getReceiver)
+                    .map(CouponMemberResponse::getId)
+                    .collect(Collectors.toList());
+            List<Long> expected = List.of(receiverId, receiverId);
+
+            assertThat(actual).isEqualTo(expected);
         }
     }
